Grogu- Isabella Robinson 11-27-2021
I wanted to focus on my favorite thing I learned while in this class and that being the 2d projects. I had done and put together  my artist piece by puting and piecing together different shapes to create my figure. I am personally a big fan of star wars and I loved the Mandalorian when it came out. More specifically grogu (baby yoda). That is what my piece is based off of. I tried to get the shape of the alien the best I could and I had put the ears up because I wanted to show off a happy version of grogu. That is my main source of inspiration.